Add secrets support, renamed existng hook.
Add encrypt-host-secrets (a user tool) and decrypt-host-secrets (a git
hook) to manage configuration files that are 'secret' and should not be
publicly viewable. Uses the host's SSH key to encrypt the files.

Renamed the existing git hook from 'post-merge' to be
'apply-host-overrides', which is more descriptive.

TODO: A simple git-hook multiplexer for post-merge.